chore(ci): replace in-house doc-links checker with lychee offline (#2308)
The `doc-links` CI gate (#638) re-implemented markdown link extraction — the
fragile part — in-house: a regex extractor with hand-rolled code-span/fence
masking. Swap it for lychee (the maintained link checker) in `--offline` mode;
markdown-native parsing gives code masking + reference/HTML-link coverage for
free (anti-NIH: use the owned tool, don't maintain a parallel one).

- .github/workflows/doc-links.yml: scan via lycheeverse/lychee-action pinned by
  SHA (v2.8.0) + lycheeVersion v0.24.2. Scope stays git-tracked-only (`git
  ls-files -z '*.md'` fed via `--files-from`), external URLs skipped (`--offline`),
  repo-absolute `/foo` links resolved via `--root-dir`. Workflow name kept
  (`doc-links`) to preserve the required-check identity.
- STRENGTHENED per ADR 0092: fail-closed on zero scope. lychee exits 0 on empty
  input (verified against lychee 0.24.2), so an explicit non-empty guard reds the
  gate on zero markdown files; `failIfEmpty: true` is the defense-in-depth backstop.
- Delete packages/pipeline-cli/src/tools/doc-links/ (tool + tests) and its
  registry wiring. Relocate the shared `findRootDir` walk (used by 6 other tools)
  to src/find-root-dir.ts so nothing dangles.

Anchor/fragment checking stays OFF (lychee default), matching the former gate.
Repo audit: no tracked doc uses the repo-absolute convention today. lychee runs
clean (0 errors) over all git-tracked .md.

Fixes #2308
