+++ packages/pipeline-cli/src/registry.ts
@@ -18,7 +18,6 @@ import {codeownersCpCommand} from "./tools/codeowners-cp/command.ts";
 import {crabboxManifestCommand} from "./tools/crabbox-manifest/command.ts";
 import {decisionsIndexCommand} from "./tools/decisions-index/command.ts";
 import {designTokenGuardCommand} from "./tools/design-token-guard/command.ts";
-import {docLinksCommand} from "./tools/doc-links/command.ts";
 import {epicLedgerCommand} from "./tools/epic-ledger/command.ts";
 import {evalHarnessCommand} from "./tools/eval-harness/command.ts";
 import {failureClassifierCommand} from "./tools/failure-classifier/command.ts";
@@ -85,7 +84,6 @@ export const registeredTools: ReadonlyArray<RegisteredTool> = [
 	spawnGuardCommand,
 	leakGuardCommand,
 	mergeQueueClassifyCommand,
-	docLinksCommand,
 	pointerGuardCommand,
 	ciRequiredCommand,
 	ghPhoenixCommand,
